Add files via upload
This game is about monkey who is trying to jump over obstacles which consist of rocks. Then the monkey needs to jump up to get bananas which basically gets them points. Once it hits a rock the game will basically stop. All of the updates are not there cause it's part one.